requirements.txt
tkinter

DateTime
os-sys
Pillow
pipwin
PyAutoGUI
pyttsx3
pywhatkit
pycopy-webbrowser
SpeechRecognition
